:recycle: refactor: gather step components
