+++ frontend/src/components/pages/events/EventForm/Steps/Review/ReviewItem.tsx
