+++ frontend/src/components/pages/events/EventForm/Steps/index.ts
@@ -1,0 +1,4 @@
+export { Info } from "./Info";
+export { Review } from "./Review";
+export { Registration } from "./Registration";
+export { TimeAndPlace } from "./TimeAndPlace";
